% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ak.R
\name{theme_ak}
\alias{theme_ak}
\title{Personal theme for ggplot2}
\usage{
theme_ak(
  base_family = "Arial",
  base_size = 12,
  grid = TRUE,
  dark = FALSE,
  alttf = FALSE,
  logo = FALSE
)
}
\arguments{
\item{grid}{Whether grid is printed or not, default=TRUE
dark Switch color scheme to dark, default=FALSE
alttf Choose alternative title font (Playfair Display)}
}
\description{
This theme relies on the IBM Plex Sans font which should be installed in advance.
For convenience, run import_ps() to install font.
}
\examples{

}
